Last Used v2 overlay
This commit it to preserve the last state of calcXtX.cpp used for overlay_v2, which was mostly used for debugging to figure out why the module output 0.

Will now be using a hard coded value for inverse of determinant until Abhiishek's approx. model works. This will allow me to create the rest of the predictor module as well as the cleaner python file for creating the driver for the linear regression IP for user friendliness.